handwritten_digit_recognition/requirements.txt
Flask
Pillow
torch
numpy
